Add Saxon-B:9.0 dependency, use in XnatPipelineLauncher
This took me a long time to figure out, and I'm not sure I got it right,
so I'm adding extra notes here. During out CNDA upgrade to XNAT 1.7, the
pipelines stopped running. One of the things that seemed to be wrong is that
the saxon library either couldn't find all the classes it needed, or it couldn't
correctly use "external libraries", which are places in the pipeline where we call
out to our own code in out own jars.

There are lots of very similar saxon libraries, and I had to figure out which to use.
The pipeline engine repo has several saxon jars checked in: saxon8-(a few variants).jar,
and saxon9.jar. The xnat-pipeline pom.xml file imported one: net.sourceforge.saxon:saxon
(with an implied "version 9" coming from the NRG parent pom). The pipeline-client
pom.xml (which I have since converted to a build.gradle) imported another:
net.sf.saxon:Saxon-B:9.0.

The script to launch the pipelines, XnatPipelineLauncher, has the three saxon8 jars on
its classpath. But since we were seeing errors that way, it seemed not correct.
The XnatPipelineLauncher class org.nrg.xnat.pipeline.client.XNATPipelineLauncher
lives in pipeline-client, which imports Saxon-B:9.0. But of all the choices, that is
the only one for which the pipeline engine did not have a jar checked in.

But that's the one I went with: Saxon-B. It seems to work now. I hope it
continues to do so.